rename HMWRK to TASK
