+++ step_definitions/createTask-steps.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const navigationLeftPage = require('../page-objects/pages/NavigationLeftPage.js');
-const homeworkListPage = require('../page-objects/pages/HMWRKHomeworkListPage.js');
-const addEditHomeworkPage = require('../page-objects/pages/HMWRKAddEditHomeworkPage.js');
+const tasksListPage = require('../page-objects/pages/TASKListPage.js/index.js');
+const addEditHomeworkPage = require('../page-objects/pages/TASKAddEditTaskPage.js');
 
 Then(/^.*clicks create-task-button on task page$/, function () {
-	return homeworkListPage.clickCreateTaskButton();
+	return tasksListPage.clickCreateTaskButton();
 });
 
 Then(/^.*puts taskname (.*) into name field$/, function (taskName) {
@@ -24,5 +24,5 @@ Then(/^.*goes to tasks page$/, function () {
 });
 
 Then(/^.*created task with name (.*) is on task page$/, async function (taskName) {
-	await homeworkListPage.isTaskVisible(taskName, true);
+	await tasksListPage.isTaskVisible(taskName, true);
 });
